twizzler::marker

Trait Invariant

pub unsafe trait Invariant { }

Indicates that a type is invariant and thus can be stored in an object.

Safety

The implementation must ensure that the type is invariant, meaning that the type must:

Be FFI safe.

Be stable in-memory (independent of architecture). This means, among other things, that the type must be fixed-width. For example, usize is not Invariant.