twizzler::object

Struct Object

pub struct Object<Base> { /* private fields */ }

Implementations

impl<Base> Object<Base>

pub fn into_tx(self) -> Result<TxObject<Base>>

Start a transaction on this object, turning this object into a transaction object handle.

Example

let obj = ObjectBuilder::new().build(12u32).unwrap();
let tx_obj = obj.into_tx().unwrap();
tx_obj.base_mut() += 1;

pub fn as_tx(&self) -> Result<TxObject<Base>>

Start a transaction on this object, creating a new transaction object handle.

Example

let obj = ObjectBuilder::new().build(12u32).unwrap();
let tx_obj = obj.as_tx().unwrap();
tx_obj.base_mut() += 1;

pub fn with_tx<R>( &mut self, f: impl FnOnce(&mut TxObject<Base>) -> Result<R>, ) -> Result<R>

Perform a transaction on this object, within the provided closure.

Example

let obj = ObjectBuilderrstarst::new().build(12u32).unwrap();
obj.with_tx(|tx| tx.base_mut() += 1).unwrap();

pub unsafe fn as_mut(&self) -> Result<MutObject<Base>>

Create a new mutable object handle from this object.

Safety

The caller must ensure that the underlying object is not changed outside of this mapping.

pub unsafe fn into_mut(self) -> Result<MutObject<Base>>

Create a new mutable object handle from this object.

Safety

The caller must ensure that the underlying object is not changed outside of this mapping.

pub unsafe fn from_handle_unchecked(handle: ObjectHandle) -> Self

pub fn from_handle(handle: ObjectHandle) -> Result<Self>

pub fn into_handle(self) -> ObjectHandle

pub unsafe fn cast<U>(self) -> Object<U>

pub fn map(id: ObjID, flags: MapFlags) -> Result<Self>

Open a new object from its ID.

The provided map flags must contain at least READ, and for stable read maps, INDIRECT. For writes, add WRITE and PERSIST.

This function checks the underlying fingerprint of the base type against the stored value and fails on mismatch to ensure type safety.

pub unsafe fn map_unchecked(id: ObjID, flags: MapFlags) -> Result<Self>

Open a new object from its ID without checking the underlying fingerprint.

Safety

This function is unsafe because it does not check the underlying fingerprint of the base type against the stored value. Use with caution.

pub fn id(&self) -> ObjID

Return the ID of the object.

pub fn update(&mut self) -> Result<()>

Update the underlying mapping of the object. This invalidates all references to the object (hence why it takes &mut self).

pub unsafe fn sync(&mut self) -> Result<()>

Sync the underlying object.

Safety

The caller must ensure that the object is not being concurrently modified by another thread, and that the memory is in a valid state to represent the object.