twizzler::object

Struct ObjectHandle

pub struct ObjectHandle(/* private fields */);

An object handle, granting access to object memory. An object handle can be in two modes:

Owning – the normal mode, which acts like an Arc, and asks the runtime to unmap when refcount hits zero.

Unsafe – internal use only, is NOT owning, but still has pointers. This is totally unsafe to use, and should not be exposed to users. But sometimes, it can be safe, and faster than cloning. … anyway, in general these have reference counting semantics, via Clone and Drop, like Arc.

Implementations

impl ObjectHandle

pub fn start(&self) -> *mut u8

Get a pointer to the start of object data.

pub fn meta(&self) -> *mut MetaInfo

Get a pointer to the metadata structure.

pub fn meta_exts(&self) -> &[MetaExt]

Get a slice of metadata extensions

pub fn find_meta_ext(&self, tag: MetaExtTag) -> Option<&MetaExt>

Find the first metadata extension matching the given tag

pub fn runtime_info(&self) -> *mut u8

Get a pointer to the runtime info.

pub fn map_flags(&self) -> MapFlags

Get map flags.

pub fn valid_len(&self) -> usize

Get the number of valid bytes after start pointer for object data.

pub fn id(&self) -> ObjID

Get the object ID.

pub fn into_raw(self) -> object_handle

Get the raw object handle.

pub fn from_raw(raw: object_handle) -> ObjectHandle

Build an object handle from raw.

pub fn cmd(&self, cmd: ObjectCmd, arg: u64) -> Result<(), TwzError>

Modify an object.

pub unsafe fn new( id: ObjID, runtime_info: *mut c_void, start: *mut c_void, meta: *mut c_void, map_flags: MapFlags, valid_len: usize, ) -> ObjectHandle

Make a new object handle.

Safety

The caller must ensure that runtime_info is a valid pointer, and points to a repr(C) struct that starts with an AtomicU64 for reference counting.