Trait twizzler::object::InitializedObject

pub trait InitializedObject {
    type Base: BaseType;

    // Required methods
    fn base(&self) -> BaseRef<'_, Self::Base>;
    fn meta(&self) -> &MetaInfo;
    fn freeze(&self) -> Result<ImmutableObject<Self::Base>, ()>;

    // Provided methods
    fn map(id: ObjID, flags: MapFlags) -> Result<Self, MapError>
       where Self: Sized { ... }
    fn resolve<T>(&self, ptr: &InvPtr<T>) -> Result<ResolvedPtr<'_, T>, ()> { ... }
}

Required Associated Types

type Base: BaseType

Required Methods

fn base(&self) -> BaseRef<'_, Self::Base>

fn meta(&self) -> &MetaInfo

fn freeze(&self) -> Result<ImmutableObject<Self::Base>, ()>

Provided Methods

fn map(id: ObjID, flags: MapFlags) -> Result<Self, MapError>where Self: Sized,

fn resolve<T>(&self, ptr: &InvPtr<T>) -> Result<ResolvedPtr<'_, T>, ()>

Resolves an invariant pointer for this object.

This function checks to ensure the passed in invariant pointer really is part of the object. It then tries to resolve the pointer according to its contents and this object’s FOT. The resulting resolved pointer does NOT implement Deref, since it may not be memory safe to do so in general (we cannot prove that noone has a mutable reference).